+++ src/test/java/uk/gov/pay/connector/charge/util/CorporateCardSurchargeCalculatorTest.java
@@ -1,8 +1,9 @@
-package uk.gov.pay.connector.common.service.charge;
+package uk.gov.pay.connector.charge.util;
 
 
 import org.junit.Test;
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
+import uk.gov.pay.connector.charge.util.CorporateCardSurchargeCalculator;
 import uk.gov.pay.connector.gateway.model.AuthCardDetails;
 import uk.gov.pay.connector.gateway.model.PayersCardType;
 import uk.gov.pay.connector.model.domain.AuthCardDetailsBuilder;
@@ -12,8 +13,8 @@ import java.util.Optional;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static uk.gov.pay.connector.common.service.charge.CorporateCardSurchargeCalculator.getCorporateCardSurchargeFor;
-import static uk.gov.pay.connector.common.service.charge.CorporateCardSurchargeCalculator.getTotalAmountFor;
+import static uk.gov.pay.connector.charge.util.CorporateCardSurchargeCalculator.getCorporateCardSurchargeFor;
+import static uk.gov.pay.connector.charge.util.CorporateCardSurchargeCalculator.getTotalAmountFor;
 
 public class CorporateCardSurchargeCalculatorTest {
     @Test
